--- src/maker.jl
@@ -1,6 +1,0 @@
-struct Vault
-    collateral::Type{Asset}
-    stabilityFee::Float64
-    LiqRatio::Float64
-    LiqFee::Float64 # Not important
-
